src/analysis/requirements.txt
# Python分析環境の依存関係

# 形態素解析
mecab-python3>=1.0.6

# 機械学習・統計
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# 自然言語処理
transformers>=4.30.0
torch>=2.0.0
fugashi>=1.2.0
unidic-lite>=1.0.8

# 可視化
matplotlib>=3.7.0
plotly>=5.14.0

